Enable SPIRAM support on no-ble configuration.
We don't normally have SPIRAM (external RAM) support on the
default ESP32 configuration, because we don't have enough IRAM.
However, the no-ble variant is not running out of IRAM.

This supports SPIRAM for the Toit heap only.  It doesn't return
SPIRAM (external RAM) for regular malloc calls, only when
requested with heap_caps_* calls.

If no SPIRAM is detected, we run in internal RAM as before.

The C compiler workaround for revision 1 ESP32 chips is enabled.